refactor(push): isolate Web Push sender adapter
